+++ lib/src/tiny_mce_editor.dart
@@ -1,2 +1,3 @@
 export 'tiny_mce_editor_stub.dart'
+    if (dart.library.js_interop) 'tiny_mce_editor_browser.dart'
     if (dart.library.html) 'tiny_mce_editor_web.dart';